+++ packages/editor/editor-core/src/__tests__/visual-regression/_utils.ts
@@ -1,11 +1,5 @@
 import { getExampleUrl } from '@atlaskit/visual-regression/helper';
-import { insertMedia as integrationInsertMedia } from '../integration/_helpers';
-import { messages as insertBlockMessages } from '../../plugins/insert-block/ui/ToolbarInsertBlock';
-import { messages as blockTypeMessages } from '../../plugins/block-type/types';
-import { messages as textFormattingMessages } from '../../plugins/text-formatting/ui/ToolbarTextFormatting';
-import { messages as advancedTextFormattingMessages } from '../../plugins/text-formatting/ui/ToolbarAdvancedTextFormatting';
-import { messages as listsMessages } from '../../plugins/lists/messages';
-import { messages as textColorMessages } from '../../plugins/text-color/ui/ToolbarTextColor';
+
 export {
   setupMediaMocksProviders,
   editable,
@@ -24,27 +18,6 @@ export const dynamicTextViewportSizes = [
   { width: 768, height: 4000 },
   { width: 1024, height: 4000 },
 ];
-
-export const resetViewport = async page => {
-  await page.setViewport({ width: DEFAULT_WIDTH, height: DEFAULT_HEIGHT });
-};
-
-export const escapeStr = (str: string) => {
-  return `concat('${str.replace(/'/g, `', "'", '`)}', '')`;
-};
-
-export const viewportSizes = [{ width: DEFAULT_WIDTH, height: DEFAULT_HEIGHT }];
-
-export const selectByTextAndClick = async ({ page, tagName, text }) => {
-  const target = await page.$x(
-    `//${tagName}[contains(text(), ${escapeStr(text)})]`,
-  );
-  if (target.length > 0) {
-    await target[0].click();
-  } else {
-    throw new Error(`Target element is not found: ${text}`);
-  }
-};
 
 // TODO: remove this gotoExample step
 export const initEditor = async (page, appearance: string) => {
@@ -146,7 +119,7 @@ function getEditorProps(appearance: Appearance) {
   return enableAllEditorProps;
 }
 
-async function mountEditor(page: any, props) {
+export async function mountEditor(page: any, props) {
   await page.evaluate(props => {
     (window as any).__mountEditor(props);
   }, props);
@@ -200,10 +173,15 @@ export const initFullPageEditorWithAdf = async (
   });
 };
 
-export const initCommentEditorWithAdf = async (page, adf: Object) => {
+export const initCommentEditorWithAdf = async (
+  page,
+  adf: Object,
+  device?: Device,
+) => {
   await initEditorWithAdf(page, {
     adf,
     appearance: Appearance.comment,
+    device,
   });
 };
 
@@ -211,290 +189,6 @@ export const clearEditor = async page => {
   await page.evaluate(() => {
     const dom = document.querySelector('.ProseMirror') as HTMLElement;
     dom.innerHTML = '<p><br /></p>';
-  });
-};
-
-export const insertBlockMenuItem = async (
-  page,
-  menuTitle,
-  tagName = 'span',
-) => {
-  const openInsertBlockMenuSelector = `[aria-label="${
-    insertBlockMessages.insertMenu.defaultMessage
-  }"]`;
-
-  await page.click(openInsertBlockMenuSelector);
-  // Do we need to wait for something here?
-  await selectByTextAndClick({ page, text: menuTitle, tagName });
-};
-
-export const insertTable = async page => {
-  await page.click(
-    `span[aria-label="${insertBlockMessages.table.defaultMessage}"]`,
-  );
-  await page.waitForSelector('table td p');
-};
-
-type CellSelectorOpts = {
-  row: number;
-  cell?: number;
-  cellType?: 'td' | 'th';
-};
-
-export const getSelectorForTableCell = ({
-  row,
-  cell,
-  cellType = 'td',
-}: CellSelectorOpts) => {
-  const rowSelector = `table tr:nth-child(${row})`;
-  if (!cell) {
-    return rowSelector;
-  }
-
-  return `${rowSelector} > ${cellType}:nth-child(${cell})`;
-};
-
-export const getSelectorForTableRow = (row: number) => {
-  return `table tr:nth-child(${row})`;
-};
-
-export const insertMenuSelector = `span[aria-label="${
-  insertBlockMessages.insertMenu.defaultMessage
-}"]`;
-
-export const advanceFormattingMenuSelector = `span[aria-label="${
-  advancedTextFormattingMessages.moreFormatting.defaultMessage
-}"]`;
-
-export const insertMenuTests = [
-  // -----------------
-  // Insert menu items
-  // -----------------
-  {
-    name: 'Quote',
-    // click selector (dropdown menu or toolbar icon)
-    clickSelector: insertMenuSelector,
-    // menu item selector - when given, it should match item inner text
-    menuItemText: blockTypeMessages.blockquote.defaultMessage,
-    // inserted node selector - wait for the node to be inserted
-    nodeSelector: 'blockquote p',
-    // is used for testing marks and typing inside content nodes
-    content: 'text',
-    // where to test
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Code snippet',
-    menuItemText: blockTypeMessages.codeblock.defaultMessage,
-    clickSelector: insertMenuSelector,
-    nodeSelector: 'div.code-block code',
-    content: 'text',
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Panel',
-    menuItemText: blockTypeMessages.panel.defaultMessage,
-    clickSelector: insertMenuSelector,
-    nodeSelector: 'div[paneltype] p',
-    content: 'text',
-    appearance: ['full-page'],
-  },
-  {
-    name: 'Divider',
-    menuItemText: insertBlockMessages.horizontalRule.defaultMessage,
-    clickSelector: insertMenuSelector,
-    nodeSelector: 'hr',
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Date',
-    menuItemText: insertBlockMessages.date.defaultMessage,
-    clickSelector: insertMenuSelector,
-    nodeSelector: 'span[timestamp]',
-    appearance: ['full-page'],
-  },
-  {
-    name: 'Columns',
-    menuItemText: insertBlockMessages.columns.defaultMessage,
-    clickSelector: insertMenuSelector,
-    nodeSelector: 'div[data-layout-type] p',
-    content: 'text',
-    appearance: ['full-page'],
-  },
-];
-
-export const toolBarItemsTests = [
-  // -----------------
-  // Toolbar items
-  // -----------------
-  {
-    name: 'Table',
-    clickSelector: `span[aria-label="${
-      insertBlockMessages.table.defaultMessage
-    }"]`,
-    nodeSelector: 'table th p',
-    content: 'text',
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Action',
-    clickSelector: `span[aria-label="${
-      insertBlockMessages.action.defaultMessage
-    }"]`,
-    nodeSelector: 'ol[data-task-list-local-id] div',
-    content: 'text',
-    appearance: ['full-page', 'message'],
-  },
-  {
-    name: 'Decision',
-    clickSelector: `span[aria-label="${
-      insertBlockMessages.decision.defaultMessage
-    }"]`,
-    nodeSelector: 'ol[data-decision-list-local-id] div',
-    content: 'text',
-    appearance: ['message'],
-  },
-  {
-    name: 'Ordered list',
-    clickSelector: `span[aria-label="${
-      listsMessages.orderedList.defaultMessage
-    }"]`,
-    nodeSelector: 'ol li p',
-    content: 'text',
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Unordered list',
-    clickSelector: `span[aria-label="${
-      listsMessages.unorderedList.defaultMessage
-    }"]`,
-    nodeSelector: 'ul li p',
-    content: 'text',
-    appearance: ['full-page', 'comment'],
-  },
-  Array.from(Array(6).keys()).map(key => {
-    const level = key + 1;
-    return {
-      name: `Heading ${level}`,
-      clickSelector: 'button[aria-haspopup="true"]',
-      menuItemText: blockTypeMessages[`heading${level}`].defaultMessage,
-      nodeSelector: `h${level}`,
-      tagName: `h${level}`,
-      content: 'text',
-      appearance: ['full-page'], // TODO add comment mode back
-      // removing comment since throwing error Node is detached from document
-    };
-  }),
-];
-
-export const baseTests = [
-  // -----------------
-  // Marks
-  // -----------------
-  ['bold', 'italic']
-    .map(k => textFormattingMessages[k].defaultMessage)
-    .map(key => ({
-      name: key,
-      clickSelector: `span[aria-label="${key}"]`,
-      nodeSelector: '.ProseMirror p',
-      content: 'text',
-      appearance: ['full-page', 'comment'],
-    })),
-  ['underline', 'strike', 'code', 'subscript', 'superscript']
-    .map(k => advancedTextFormattingMessages[k].defaultMessage)
-    .map(key => ({
-      name: key,
-      menuItemText: key,
-      clickSelector: advanceFormattingMenuSelector,
-      nodeSelector: '.ProseMirror p',
-      content: 'text',
-      appearance: ['full-page', 'comment'],
-    })),
-  // TODO run this after the fix for 'Light grey' on master
-  // Array.from(colorPalette.values()).map(key => {
-  //   return {
-  //     name: `Text color: ${key}`,
-  //     clickSelector: `span[aria-label="${
-  //       textColorMessages.textColor.defaultMessage
-  //     }"]`,
-  //     menuItemSelector: `button[title="${key}"]`,
-  //     nodeSelector: '.ProseMirror p',
-  //     content: 'text',
-  //     appearance: ['full-page', 'comment'],
-  //   };
-  // }),
-];
-
-const dropdowns = [
-  // -----------------
-  // Dropdowns
-  // -----------------
-  {
-    name: 'Normal text dropdown',
-    clickSelector: 'button[aria-haspopup="true"]',
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Insert menu',
-    clickSelector: insertMenuSelector,
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Advance formatting menu',
-    clickSelector: advanceFormattingMenuSelector,
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Text color picker',
-    clickSelector: `span[aria-label="${
-      textColorMessages.textColor.defaultMessage
-    }"]`,
-    appearance: ['full-page', 'comment'],
-  },
-  {
-    name: 'Mention picker',
-    clickSelector: `span[aria-label="${
-      insertBlockMessages.mention.defaultMessage
-    }"]`,
-    nodeSelector: 'span[data-type-ahead-query]',
-    appearance: ['full-page', 'comment', 'message'],
-  },
-  {
-    name: 'Hyperlink Recent Search',
-    clickSelector: `span[aria-label="${
-      insertBlockMessages.link.defaultMessage
-    }"]`,
-    appearance: ['full-page', 'comment'],
-  },
-];
-
-// group tests by appearances
-export const testsByAppearance = {};
-
-const addToAppearance = test => {
-  test.appearance.forEach(appearance => {
-    if (!testsByAppearance[appearance]) {
-      testsByAppearance[appearance] = [];
-    }
-    testsByAppearance[appearance].push(test);
-  });
-};
-
-export const setTests = forInput => {
-  let testArr: any[] = baseTests;
-  if (forInput === 'insertMenu') {
-    testArr = insertMenuTests;
-  } else if (forInput === 'toolbar') {
-    testArr = toolBarItemsTests;
-  } else if (forInput === 'dropdown') {
-    testArr = dropdowns;
-  }
-  testArr.forEach(test => {
-    if (Array.isArray(test)) {
-      test.forEach(addToAppearance);
-    } else {
-      addToAppearance(test);
-    }
   });
 };
 
@@ -524,28 +218,4 @@ export const snapshot = async (
     // @ts-ignore
     expect(image).toMatchProdImageSnapshot();
   }
-};
-
-export const insertMedia = async (page, filenames = ['one.svg']) => {
-  // We need to wrap this as the xpath selector used in integration tests
-  // isnt valid in puppeteer
-  await integrationInsertMedia(page, filenames, 'div[aria-label="%s"]');
-};
-
-// Execute the click using page.evaluate
-// There appears to be a bug in Puppeteer which causes the
-// "Node is either not visible or not an HTMLElement" error.
-// https://product-fabric.atlassian.net/browse/ED-5688
-export const evaluateClick = (page, selector) => {
-  return page.evaluate(selector => {
-    document.querySelector(selector).click();
-  }, selector);
-};
-
-export const getBoundingRect = async (page, selector) => {
-  return await page.evaluate(selector => {
-    const element = document.querySelector(selector);
-    const { x, y, width, height } = element.getBoundingClientRect();
-    return { left: x, top: y, width, height, id: element.id };
-  }, selector);
 };