+++ packages/editor/editor-core/src/plugins/hyperlink/pm-plugins/keymap.ts
@@ -17,16 +17,14 @@ export function createKeymapPlugin(
 ): Plugin | undefined {
   const list = {};
 
-  if (props.appearance !== 'message') {
-    keymaps.bindKeymapWithCommand(
-      keymaps.addLink.common!,
-      trackAndInvoke(
-        'atlassian.editor.format.hyperlink.keyboard',
-        showLinkToolbar(INPUT_METHOD.SHORTCUT),
-      ),
-      list,
-    );
-  }
+  keymaps.bindKeymapWithCommand(
+    keymaps.addLink.common!,
+    trackAndInvoke(
+      'atlassian.editor.format.hyperlink.keyboard',
+      showLinkToolbar(INPUT_METHOD.SHORTCUT),
+    ),
+    list,
+  );
 
   keymaps.bindKeymapWithCommand(
     keymaps.enter.common!,