+++ packages/editor/editor-core/src/utils/mark.ts
@@ -70,3 +70,29 @@ export const removeBlockMarks = (
   });
   return blockMarksExists ? tr : undefined;
 };
+
+/**
+ * Removes marks from nodes in the current selection that are not supported
+ */
+export const sanitizeSelectionMarks = (
+  state: EditorState,
+): Transaction | undefined => {
+  let tr;
+  const { $from, $to } = state.tr.selection;
+  const nodeType = $from.node().type;
+  state.doc.nodesBetween($from.pos, $to.pos, (node, pos) => {
+    const nodePos = state.doc.resolve(pos);
+    node.marks.forEach(mark => {
+      if (!nodeType.allowsMarkType(mark.type)) {
+        const filteredMarks = node.marks.filter(m => m.type !== mark.type);
+        tr = (tr || state.tr).setNodeMarkup(
+          nodePos.pos,
+          undefined,
+          node.attrs,
+          filteredMarks,
+        );
+      }
+    });
+  });
+  return tr;
+};