+++ packages/editor/editor-wikimarkup-transformer/src/index.ts
@@ -4,6 +4,7 @@ import { Node as PMNode, Schema } from 'prosemirror-model';
 import { encode } from './encoder';
 import AbstractTree from './parser/abstract-tree';
 import { Context } from './parser/tokenize';
+import { buildIssueKeyRegex } from './parser/tokenize/issue-key';
 
 export class WikiMarkupTransformer implements Transformer<string> {
   private schema: Schema;
@@ -19,7 +20,16 @@ export class WikiMarkupTransformer implements Transformer<string> {
   parse(wikiMarkup: string, context?: Context): PMNode {
     const tree = new AbstractTree(this.schema, wikiMarkup);
 
-    return tree.getProseMirrorModel(context ? context : {});
+    return tree.getProseMirrorModel(this.buildContext(context));
+  }
+
+  private buildContext(context?: Context): Context {
+    return context
+      ? {
+          ...context,
+          issueKeyRegex: buildIssueKeyRegex(context.inlineCardConversion),
+        }
+      : {};
   }
 }