+++ packages/navigation/atlassian-switcher/examples/helpers/index.tsx
@@ -1,5 +1,8 @@
 import React, { Component } from 'react';
 import SettingsIcon from '@atlaskit/icon/glyph/settings';
+import { AnalyticsListener } from '@atlaskit/analytics-next';
+import { UIAnalyticsEventInterface } from '@atlaskit/analytics-next-types';
+
 import {
   SwitcherWrapper,
   SwitcherItem,
@@ -9,7 +12,7 @@ import {
 } from '../../src/primitives';
 import MockProvider from './mock-provider';
 
-class Switcher extends Component {
+export class Switcher extends Component {
   render() {
     return (
       <MockProvider>
@@ -18,20 +21,20 @@ class Switcher extends Component {
             <Skeleton />
           ) : (
             <SwitcherWrapper>
-              <Section isAdmin title="First Section">
-                <SwitcherItem icon={SettingsIcon}>
+              <Section sectionId="first-section" isAdmin title="First Section">
+                <SwitcherItem icon={SettingsIcon} href="/">
                   {`${data && data.data} First Item`}
                 </SwitcherItem>
                 <SwitcherItem>{`${data &&
                   data.data} Second Item`}</SwitcherItem>
                 <SwitcherItem>{`${data && data.data} Third Item`}</SwitcherItem>
               </Section>
-              <Section title="Second Section">
+              <Section sectionId="second-section" title="Second Section">
                 <SwitcherItem>First Item</SwitcherItem>
                 <SwitcherItem>Second Item</SwitcherItem>
                 <SwitcherItem>Third Item</SwitcherItem>
               </Section>
-              <Section title="Third Section">
+              <Section sectionId="third-section" title="Third Section">
                 <SwitcherItem>First Item</SwitcherItem>
                 <SwitcherItem>Second Item</SwitcherItem>
                 <SwitcherItem>Third Item</SwitcherItem>
@@ -54,4 +57,25 @@ class Switcher extends Component {
   }
 }
 
-export default Switcher;
+const onAnalyticsEvent = (event: UIAnalyticsEventInterface, channel) => {
+  // tslint:disable-next-line:no-console
+  console.log(
+    `AnalyticsEvent(${channel})\n\tpayload=%o\n\tcontext=%o`,
+    event.payload,
+    event.context,
+  );
+};
+
+export const AnalyticsLogger = ({ children }) => {
+  return (
+    <AnalyticsListener channel="*" onEvent={onAnalyticsEvent}>
+      {children}
+    </AnalyticsListener>
+  );
+};
+
+export const withAnalyticsLogger = WrappedComponent => props => (
+  <AnalyticsLogger>
+    <WrappedComponent {...props} />
+  </AnalyticsLogger>
+);