+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 3.1.0
+- [minor] [8709be280f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8709be280f):
+
+  - Add issue key token to convert into inline cards (Jira Smart Cards)
+
+## 3.0.3
+- Updated dependencies [4a84fc40e0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4a84fc40e0):
+  - @atlaskit/editor-json-transformer@4.1.12
+  - @atlaskit/editor-test-helpers@7.0.1
+  - @atlaskit/editor-core@101.0.0
+  - @atlaskit/renderer@40.0.0
+
 ## 3.0.2
 - Updated dependencies [4af5bd2a58](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4af5bd2a58):
   - @atlaskit/editor-json-transformer@4.1.11