+++ packages/elements/task-decision/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/task-decision
 
+## 11.2.2
+- Updated dependencies [4a84fc40e0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4a84fc40e0):
+  - @atlaskit/renderer@40.0.0
+
 ## 11.2.1
 - Updated dependencies [fc6164c8c2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fc6164c8c2):
   - @atlaskit/editor-common@32.0.0